requirements.txt
Django==5.1.2
djangorestframework==3.15.2
gunicorn==22.0.0
